restructuring openapi_go_generator
