+++ plugins/go/examples/bookstore/Makefile
@@ -1,10 +1,10 @@
 
 build:	
-	go install github.com/googleapis/openapi-compiler/plugins/go/openapi_go_generator/template_compiler
-	pushd ../../openapi_go_generator; template_compiler; popd
 	go install github.com/googleapis/openapi-compiler/openapic
+	go install github.com/googleapis/openapi-compiler/plugins/go/openapi_go_generator/encode_templates
+	go generate github.com/googleapis/openapi-compiler/plugins/go/openapi_go_generator
 	go install github.com/googleapis/openapi-compiler/plugins/go/openapi_go_generator
-	rm $(GOPATH)/bin/openapi_go_client $(GOPATH)/bin/openapi_go_service
+	rm -f $(GOPATH)/bin/openapi_go_client $(GOPATH)/bin/openapi_go_service
 	ln -s $(GOPATH)/bin/openapi_go_generator $(GOPATH)/bin/openapi_go_client
 	ln -s $(GOPATH)/bin/openapi_go_generator $(GOPATH)/bin/openapi_go_service
