+++ plugins/go/openapi_go_generator/main.go
@@ -12,21 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+//go:generate encode_templates
+
 package main
 
 import (
-	"bytes"
-	"encoding/base64"
-	"fmt"
 	"io/ioutil"
 	"log"
 	"os"
-	"os/exec"
-	"path"
-	"path/filepath"
-	"runtime"
-	"strings"
-	"text/template"
 
 	"github.com/golang/protobuf/proto"
 
@@ -34,364 +27,9 @@ import (
 	plugins "github.com/googleapis/openapi-compiler/plugins"
 )
 
-type ServiceType struct {
-	Name   string
-	Fields []*ServiceTypeField
-}
-
-func (s *ServiceType) hasFieldNamed(name string) bool {
-	for _, f := range s.Fields {
-		if f.Name == name {
-			return true
-		}
-	}
-	return false
-}
-
-type ServiceTypeField struct {
-	Name     string
-	Type     string
-	JSONName string
-	Position string // "body", "header", "formdata", "query", or "path"
-}
-
-type ServiceMethod struct {
-	Name               string
-	Path               string
-	Method             string
-	Description        string
-	HandlerName        string
-	ProcessorName      string
-	ClientName         string
-	ResultTypeName     string
-	ParametersTypeName string
-	ResponsesTypeName  string
-	ParametersType     *ServiceType
-	ResponsesType      *ServiceType
-}
-
-type ServiceRenderer struct {
-	Templates map[string]*template.Template
-
-	Name    string
-	Package string
-	Types   []*ServiceType
-	Methods []*ServiceMethod
-}
-
-func NewServiceRenderer(document *openapi.Document, packageName string) (renderer *ServiceRenderer, err error) {
-	renderer = &ServiceRenderer{}
-	renderer.Name = document.Info.Title
-	renderer.Package = packageName
-	err = renderer.loadService(document)
-	if err != nil {
-		return nil, err
-	}
-	return renderer, nil
-}
-
-// template support functions
-
-func hasOKField(s *ServiceType) bool {
-	return s.hasFieldNamed("OK")
-}
-
-func parameterList(m *ServiceMethod) string {
-	result := ""
-	for i, field := range m.ParametersType.Fields {
-		if i > 0 {
-			result += ", "
-		}
-		result += field.JSONName + " " + field.Type
-	}
-	return result
-}
-
-func bodyParameterName(m *ServiceMethod) string {
-	for _, field := range m.ParametersType.Fields {
-		if field.Position == "body" {
-			return field.JSONName
-		}
-	}
-	return ""
-}
-
-func bodyParameterFieldName(m *ServiceMethod) string {
-	for _, field := range m.ParametersType.Fields {
-		if field.Position == "body" {
-			return field.Name
-		}
-	}
-	return ""
-}
-
-// instantiate templates
-
-func (renderer *ServiceRenderer) loadTemplates(files map[string]string) (err error) {
-	renderer.Templates = make(map[string]*template.Template, 0)
-
-	funcMap := template.FuncMap{
-		"HASOK":                  hasOKField,
-		"parameterList":          parameterList,
-		"bodyParameterName":      bodyParameterName,
-		"bodyParameterFieldName": bodyParameterFieldName,
-	}
-
-	for filename, encoding := range files {
-
-		templateData, err := base64.StdEncoding.DecodeString(encoding)
-		if err != nil {
-			return err
-		}
-
-		templateString := string(templateData)
-		t, err := template.New(filename).Funcs(funcMap).Parse(templateString)
-		if err != nil {
-			log.Printf("ERROR: %+v", err)
-			return err
-		} else {
-			renderer.Templates[filename] = t
-		}
-	}
-	return err
-}
-
-func (renderer *ServiceRenderer) loadServiceTypeFromParameters(name string, parameters []*openapi.ParametersItem) (t *ServiceType, err error) {
-	t = &ServiceType{}
-	t.Fields = make([]*ServiceTypeField, 0)
-	for _, parametersItem := range parameters {
-		var f ServiceTypeField
-		f.Type = fmt.Sprintf("%+v", parametersItem)
-		parameter := parametersItem.GetParameter()
-		if parameter != nil {
-			bodyParameter := parameter.GetBodyParameter()
-			if bodyParameter != nil {
-				f.Name = bodyParameter.Name
-				if bodyParameter.Schema != nil {
-					f.Type = typeForSchema(bodyParameter.Schema)
-					f.Position = "body"
-				}
-			}
-			nonBodyParameter := parameter.GetNonBodyParameter()
-			if nonBodyParameter != nil {
-				headerParameter := nonBodyParameter.GetHeaderParameterSubSchema()
-				if headerParameter != nil {
-					f.Name = headerParameter.Name
-					f.Position = "header"
-				}
-				formDataParameter := nonBodyParameter.GetFormDataParameterSubSchema()
-				if formDataParameter != nil {
-					f.Name = formDataParameter.Name
-					f.Position = "formdata"
-				}
-				queryParameter := nonBodyParameter.GetQueryParameterSubSchema()
-				if queryParameter != nil {
-					f.Name = queryParameter.Name
-					f.Position = "query"
-				}
-				pathParameter := nonBodyParameter.GetPathParameterSubSchema()
-				if pathParameter != nil {
-					f.Name = pathParameter.Name
-					f.Position = "path"
-					f.Type = typeForName(pathParameter.Type, pathParameter.Format)
-				}
-			}
-			f.JSONName = f.Name
-			f.Name = strings.Title(f.Name)
-			t.Fields = append(t.Fields, &f)
-		}
-	}
-	t.Name = name
-	renderer.Types = append(renderer.Types, t)
-	return t, err
-}
-
-func propertyNameForResponseCode(code string) string {
-	if code == "200" {
-		return "OK"
-	}
-	return code
-}
-
-func (renderer *ServiceRenderer) loadServiceTypeFromResponses(m *ServiceMethod, name string, responses *openapi.Responses) (t *ServiceType, err error) {
-	t = &ServiceType{}
-	t.Fields = make([]*ServiceTypeField, 0)
-
-	for _, responseCode := range responses.ResponseCode {
-		var f ServiceTypeField
-		f.Name = propertyNameForResponseCode(responseCode.Name)
-		f.JSONName = ""
-		response := responseCode.Value.GetResponse()
-		if response != nil && response.Schema != nil && response.Schema.GetSchema() != nil {
-			f.Type = "*" + typeForSchema(response.Schema.GetSchema())
-			t.Fields = append(t.Fields, &f)
-			if f.Name == "OK" {
-				m.ResultTypeName = typeForSchema(response.Schema.GetSchema())
-			}
-		}
-	}
-
-	t.Name = name
-	renderer.Types = append(renderer.Types, t)
-	return t, err
-}
-
-func (renderer *ServiceRenderer) loadOperation(op *openapi.Operation, method string, path string) (err error) {
-	var m ServiceMethod
-	m.Name = strings.Title(op.OperationId)
-	m.Path = path
-	m.Method = method
-	m.HandlerName = "handle" + m.Name
-	m.ProcessorName = "process" + m.Name
-	m.ClientName = m.Name
-	m.ParametersTypeName = m.Name + "Parameters"
-	m.ResponsesTypeName = m.Name + "Responses"
-	m.ParametersType, err = renderer.loadServiceTypeFromParameters(m.ParametersTypeName, op.Parameters)
-	m.ResponsesType, err = renderer.loadServiceTypeFromResponses(&m, m.ResponsesTypeName, op.Responses)
-	renderer.Methods = append(renderer.Methods, &m)
-	return err
-}
-
-// preprocess the types and methods of the API
-func (renderer *ServiceRenderer) loadService(document *openapi.Document) (err error) {
-	// collect service type descriptions
-	renderer.Types = make([]*ServiceType, 0)
-	for _, pair := range document.Definitions.AdditionalProperties {
-		var t ServiceType
-		t.Fields = make([]*ServiceTypeField, 0)
-		schema := pair.Value
-		for _, pair2 := range schema.Properties.AdditionalProperties {
-			var f ServiceTypeField
-			f.Name = strings.Title(pair2.Name)
-			f.Type = typeForSchema(pair2.Value)
-			f.JSONName = pair2.Name
-			t.Fields = append(t.Fields, &f)
-		}
-		t.Name = strings.Title(filteredTypeName(pair.Name))
-		renderer.Types = append(renderer.Types, &t)
-	}
-	// collect service method descriptions
-	renderer.Methods = make([]*ServiceMethod, 0)
-	for _, pair := range document.Paths.Path {
-		v := pair.Value
-		if v.Get != nil {
-			renderer.loadOperation(v.Get, "GET", pair.Name)
-		}
-		if v.Post != nil {
-			renderer.loadOperation(v.Post, "POST", pair.Name)
-		}
-		if v.Put != nil {
-			renderer.loadOperation(v.Put, "PUT", pair.Name)
-		}
-		if v.Delete != nil {
-			renderer.loadOperation(v.Delete, "DELETE", pair.Name)
-		}
-	}
-	return err
-}
-
-func filteredTypeName(typeName string) (name string) {
-	// first take the last path segment
-	parts := strings.Split(typeName, "/")
-	name = parts[len(parts)-1]
-	// then take the last part of a dotted name
-	parts = strings.Split(name, ".")
-	name = parts[len(parts)-1]
-	return name
-}
-
-func typeForName(name string, format string) (typeName string) {
-	switch name {
-	case "integer":
-		if format == "int32" {
-			return "int32"
-		} else if format == "int64" {
-			return "int64"
-		} else {
-			return "int32"
-		}
-	default:
-		return name
-	}
-}
-
-func typeForSchema(schema *openapi.Schema) (typeName string) {
-	ref := schema.XRef
-	if ref != "" {
-		return typeForRef(ref)
-	}
-	if schema.Type != nil {
-		types := schema.Type.Value
-		if len(types) == 1 && types[0] == "string" {
-			return "string"
-		}
-		if len(types) == 1 && types[0] == "array" && schema.Items != nil {
-			// we have an array.., but of what?
-			items := schema.Items.Schema
-			if len(items) == 1 && items[0].XRef != "" {
-				return "[]" + typeForRef(items[0].XRef)
-			}
-		}
-	}
-	// this function is incomplete... so return a string representing anything that we don't handle
-	return fmt.Sprintf("%v", schema)
-}
-
-func typeForRef(ref string) (typeName string) {
-	return strings.Title(path.Base(ref))
-}
-
-func gofmt(inputBytes []byte) (outputBytes []byte, err error) {
-	cmd := exec.Command(runtime.GOROOT() + "/bin/gofmt")
-	input, _ := cmd.StdinPipe()
-	output, _ := cmd.StdoutPipe()
-	cmderr, _ := cmd.StderrPipe()
-	err = cmd.Start()
-	if err != nil {
-		return
-	}
-	input.Write(inputBytes)
-	input.Close()
-	errors, err := ioutil.ReadAll(cmderr)
-	if len(errors) > 0 {
-		return inputBytes, nil
-	} else {
-		outputBytes, err = ioutil.ReadAll(output)
-	}
-	return
-}
-
-func (renderer *ServiceRenderer) GenerateApp(response *plugins.PluginResponse, files []string) (err error) {
-
-	for _, filename := range files {
-		file := &plugins.File{}
-		file.Name = filename
-		f := new(bytes.Buffer)
-		t := renderer.Templates[filename]
-		err = t.Execute(f, struct {
-			Renderer *ServiceRenderer
-		}{
-			renderer,
-		})
-		if err != nil {
-			response.Text = append(response.Text, fmt.Sprintf("ERROR %v", err))
-		}
-		inputBytes := f.Bytes()
-		if filepath.Ext(file.Name) == ".go" {
-			file.Data, err = gofmt(inputBytes)
-		} else {
-			file.Data = inputBytes
-		}
-		response.File = append(response.File, file)
-	}
-	return
-}
-
 // run the plugin
 
 func main() {
-
-	log.Printf("RUNNING %s", os.Args[0])
 
 	var files []string
 	switch os.Args[0] {
@@ -418,42 +56,39 @@ func main() {
 		document := &openapi.Document{}
 		err = proto.Unmarshal(wrapper.Value, document)
 		if err != nil {
-			panic(err)
+			log.Printf("ERROR %v", err)
+			os.Exit(1)
 		}
 
-		name := wrapper.Name
-		version := wrapper.Version
+		invocation := os.Args[0]
+
 		parameters := request.Parameters
-
-		log.Printf("Reading %s (%s)", name, version)
-
 		packageName := "main"
 		for _, parameter := range parameters {
+			invocation += " " + parameter.Name + "=" + parameter.Value
 			if parameter.Name == "package" {
 				packageName = parameter.Value
 			}
 		}
 
+		log.Printf("Running %s", invocation)
+
 		renderer, err := NewServiceRenderer(document, packageName)
 		if err != nil {
 			log.Printf("ERROR %v", err)
-			return
+			os.Exit(1)
 		}
 
 		err = renderer.loadTemplates(templates())
 		if err != nil {
 			log.Printf("ERROR %v", err)
-			return
+			os.Exit(1)
 		}
 
+		err = renderer.Generate(response, files)
 		if err != nil {
 			log.Printf("ERROR %v", err)
-			return
-		}
-		err = renderer.GenerateApp(response, files)
-		if err != nil {
-			log.Printf("ERROR %v", err)
-			return
+			os.Exit(1)
 		}
 	}
