+++ plugins/go/openapi_go_generator/templates.go
@@ -1,0 +1,11 @@
+package main
+
+func templates() map[string]string {
+	return map[string]string{ 
+        "app.go": "cGFja2FnZSB7ey5SZW5kZXJlci5QYWNrYWdlfX0KCmltcG9ydCAoCgkiZW5jb2RpbmcvanNvbiIKCSJmbXQiCgkibG9nIgoJIm5ldC9odHRwIgoJInN0cmNvbnYiCgoJImdpdGh1Yi5jb20vZ29yaWxsYS9tdXgiCikKCmZ1bmMgaW50VmFsdWUocyBzdHJpbmcpICh2IGludDY0KSB7Cgl2LCBfID0gc3RyY29udi5QYXJzZUludChzLCAxMCwgNjQpCglyZXR1cm4gdgp9CgovLyBoYW5kbGVycwp7e3JhbmdlIC5SZW5kZXJlci5NZXRob2RzfX0KCi8vIHt7LkRlc2NyaXB0aW9ufX0KZnVuYyB7ey5IYW5kbGVyTmFtZX19KHcgaHR0cC5SZXNwb25zZVdyaXRlciwgciAqaHR0cC5SZXF1ZXN0KSB7Cglsb2cuUHJpbnRmKCJ7ey5OYW1lfX0iKTsKCXZhciBlcnIgZXJyb3IKCS8vIGluc3RhbnRpYXRlIHRoZSBwYXJhbWV0ZXJzIHN0cnVjdHVyZQoJdmFyIHBhcmFtZXRlcnMge3suUGFyYW1ldGVyc1R5cGVOYW1lfX0KCXt7aWYgZXEgLk1ldGhvZCAiUE9TVCJ9fS8vIGRlc2VyaWFsaXplIHJlcXVlc3QgZnJvbSBwb3N0IGRhdGEKCWRlY29kZXIgOj0ganNvbi5OZXdEZWNvZGVyKHIuQm9keSkKCWVyciA9IGRlY29kZXIuRGVjb2RlKCZwYXJhbWV0ZXJzLnt7Ym9keVBhcmFtZXRlckZpZWxkTmFtZSAufX0pCglpZiBlcnIgIT0gbmlsIHsKCQlmbXQuRnByaW50Zih3LCAiRVJST1I6ICV2IiwgZXJyKQoJCXJldHVybgoJfQoJbG9nLlByaW50ZigiUEFSQU1FVEVSUyAlK3YiLCBwYXJhbWV0ZXJzKXt7ZW5kfX0KCQoJLy8gZ2V0IHJlcXVlc3QgZmllbGRzIGluIHBhdGggYW5kIHF1ZXJ5IHBhcmFtZXRlcnMKCXZhcnMgOj0gbXV4LlZhcnMocikKCXIuUGFyc2VGb3JtKCkKCWxvZy5QcmludGYoIk1VWCBWQVJTICUrdiIsIHZhcnMpCglsb2cuUHJpbnRmKCJRVUVSWSBQQVJNUyAlK3YiLCByLkZvcm0pCgl7e3JhbmdlIC5QYXJhbWV0ZXJzVHlwZS5GaWVsZHN9fQkKCQoJe3tpZiBlcSAuUG9zaXRpb24gInBhdGgifX0KCWlmIHZhbHVlLCBvayA6PSB2YXJzWyJ7ey5KU09OTmFtZX19Il07IG9rIHsKCQlwYXJhbWV0ZXJzLnt7Lk5hbWV9fSA9IGludFZhbHVlKHZhbHVlKQoJfQoJe3tlbmR9fQoJCgl7e2lmIGVxIC5Qb3NpdGlvbiAiZm9ybWRhdGEifX0KCWlmIGxlbihyLkZvcm1bInt7LkpTT05OYW1lfX0iXSkgPiAwIHsKCQlwYXJhbWV0ZXJzLnt7Lk5hbWV9fSA9IGludFZhbHVlKHIuRm9ybVsie3suSlNPTk5hbWV9fSJdWzBdKQoJfQoJe3tlbmR9fQoJCgl7e2VuZH19CgkJCgkvLyBpbnN0YW50aWF0ZSB0aGUgcmVzcG9uc2VzIHN0cnVjdHVyZQoJdmFyIHJlc3BvbnNlcyB7ey5SZXNwb25zZXNUeXBlTmFtZX19CgkKCS8vIGNhbGwgdGhlIHByb2Nlc3NvcgoJZXJyID0ge3suUHJvY2Vzc29yTmFtZX19KCZwYXJhbWV0ZXJzLCAmcmVzcG9uc2VzKQoJCglpZiBlcnIgPT0gbmlsIHsKCQl7eyBpZiAuUmVzcG9uc2VzVHlwZSB8IEhBU09LIH19CgkJLy8gd3JpdGUgdGhlIG5vcm1hbCByZXNwb25zZQkJCgkJZW5jb2RlciA6PSBqc29uLk5ld0VuY29kZXIodykKCQllbmNvZGVyLkVuY29kZShyZXNwb25zZXMuT0spCgkJe3sgZW5kIH19Cgl9IGVsc2UgewoJCS8vIHdyaXRlIGFuIGVycm9yIHJlc3BvbnNlLi4uCgkJZm10LkZwcmludGYodywgIkVSUk9SOiAldiIsIGVycikKCX0KfQp7e2VuZH19CgpmdW5jIGluaXQgKCkgewoJdmFyIHJvdXRlciA9IG11eC5OZXdSb3V0ZXIoKXt7cmFuZ2UgLlJlbmRlcmVyLk1ldGhvZHN9fQoJcm91dGVyLkhhbmRsZUZ1bmMoInt7LlBhdGh9fSIsIHt7LkhhbmRsZXJOYW1lfX0pLk1ldGhvZHMoInt7Lk1ldGhvZH19Iil7e2VuZH19CiAgICBodHRwLkhhbmRsZSgiLyIsIHJvdXRlcikKICAgIGluaXRpYWxpemVfc2VydmljZSgpCn0KCmZ1bmMgbWFpbiAoKSB7CglodHRwLkxpc3RlbkFuZFNlcnZlKCI6ODA4MCIsIG5pbCkKfQ==",
+        "app.yaml": "YXBwbGljYXRpb246IHt7LlJlbmRlcmVyLk5hbWV9fQp2ZXJzaW9uOiAxCnJ1bnRpbWU6IGdvCmFwaV92ZXJzaW9uOiBnbzEKaGFuZGxlcnM6Ci0gdXJsOiAvLioKICBzY3JpcHQ6IF9nb19hcHAKLSB1cmw6IC8KICBzdGF0aWNfZGlyOiBzdGF0aWM=",
+        "client.go": "cGFja2FnZSB7ey5SZW5kZXJlci5QYWNrYWdlfX0KCmltcG9ydCAoCiAgImJ5dGVzIgogICJlbmNvZGluZy9qc29uIgogICJmbXQiCiAgIm5ldC9odHRwIgogICJzdHJpbmdzIgopCiAgCmNvbnN0IHNlcnZpY2UgPSAiaHR0cDovL2xvY2FsaG9zdDo4MDgwIgoKdHlwZSBDbGllbnQgc3RydWN0IHsKCVNlcnZpY2Ugc3RyaW5nCn0gCgpmdW5jIE5ld0NsaWVudChzZXJ2aWNlIHN0cmluZykgKkNsaWVudCB7CgljbGllbnQgOj0gJkNsaWVudHt9CgljbGllbnQuU2VydmljZSA9IHNlcnZpY2UKCXJldHVybiBjbGllbnQKfQoKe3tyYW5nZSAuUmVuZGVyZXIuTWV0aG9kc319Cnt7aWYgZXEgLlJlc3VsdFR5cGVOYW1lICIifX0KZnVuYyAoY2xpZW50ICpDbGllbnQpIHt7LkNsaWVudE5hbWV9fSh7e3BhcmFtZXRlckxpc3QgLn19KSAoZXJyIGVycm9yKSB7Cnt7ZWxzZX19CmZ1bmMgKGNsaWVudCAqQ2xpZW50KSB7ey5DbGllbnROYW1lfX0oe3twYXJhbWV0ZXJMaXN0IC59fSkgKHJlc3VsdCAqe3suUmVzdWx0VHlwZU5hbWV9fSwgZXJyIGVycm9yKSB7Cnt7ZW5kfX0KCXBhdGggOj0gY2xpZW50LlNlcnZpY2UgKyAie3suUGF0aH19IgoJCgl7e3JhbmdlIC5QYXJhbWV0ZXJzVHlwZS5GaWVsZHN9fQkKCXt7aWYgZXEgLlBvc2l0aW9uICJwYXRoIn19CglwYXRoID0gc3RyaW5ncy5SZXBsYWNlKHBhdGgsICJ7IiArICJ7ey5KU09OTmFtZX19IiArICJ9IiwgZm10LlNwcmludGYoIiV2Iiwge3suSlNPTk5hbWV9fSksIDEpCgl7e2VuZH19Cgl7e2VuZH19CgkKCXt7aWYgZXEgLk1ldGhvZCAiUE9TVCJ9fQoJYm9keSA6PSBuZXcoYnl0ZXMuQnVmZmVyKQoJanNvbi5OZXdFbmNvZGVyKGJvZHkpLkVuY29kZSh7e2JvZHlQYXJhbWV0ZXJOYW1lIC59fSkKCXJlcSwgZXJyIDo9IGh0dHAuTmV3UmVxdWVzdCgie3suTWV0aG9kfX0iLCBwYXRoLCBib2R5KQoJe3tlbHNlfX0KCXJlcSwgZXJyIDo9IGh0dHAuTmV3UmVxdWVzdCgie3suTWV0aG9kfX0iLCBwYXRoLCBuaWwpCgl7e2VuZH19CgkKCWlmIGVyciAhPSBuaWwgewoJCXJldHVybgoJfQoJcmVzcCwgZXJyIDo9IGh0dHAuRGVmYXVsdENsaWVudC5EbyhyZXEpCglpZiBlcnIgIT0gbmlsIHsKCQlyZXR1cm4gCgl9CglkZWZlciByZXNwLkJvZHkuQ2xvc2UoKQoJe3tpZiBuZSAuUmVzdWx0VHlwZU5hbWUgIiJ9fQoJZGVjb2RlciA6PSBqc29uLk5ld0RlY29kZXIocmVzcC5Cb2R5KQoJcmVzdWx0ID0gJnt7LlJlc3VsdFR5cGVOYW1lfX17fQoJZGVjb2Rlci5EZWNvZGUocmVzdWx0KQoJe3tlbmR9fQoJcmV0dXJuICAKfQp7e2VuZH19Cg==",
+        "service.go": "cGFja2FnZSB7ey5SZW5kZXJlci5QYWNrYWdlfX0KCmltcG9ydCAoCgkibG9nIgopCgpmdW5jIGluaXRpYWxpemVfc2VydmljZSgpIHsKfQoKe3tyYW5nZSAuUmVuZGVyZXIuTWV0aG9kc319CmZ1bmMge3suUHJvY2Vzc29yTmFtZX19KHBhcmFtZXRlcnMgKnt7LlBhcmFtZXRlcnNUeXBlTmFtZX19LCByZXNwb25zZXMgKnt7LlJlc3BvbnNlc1R5cGVOYW1lfX0pIChlcnIgZXJyb3IpIHsKCWxvZy5QcmludGYoInt7Lk5hbWV9fSAlK3YiLCBwYXJhbWV0ZXJzKQoJLy8gQ09NUExFVEUgVEhJUyBNQU5VQUxMWQoJcmV0dXJuIG5pbAp9Cnt7ZW5kfX0=",
+        "types.go": "cGFja2FnZSB7ey5SZW5kZXJlci5QYWNrYWdlfX0KCmltcG9ydCAoCikKCi8vIHR5cGVzCnt7cmFuZ2UgLlJlbmRlcmVyLlR5cGVzfX0KdHlwZSB7ey5OYW1lfX0gc3RydWN0IHsge3tyYW5nZSAuRmllbGRzfX0KICB7ey5OYW1lfX0ge3suVHlwZX19e3tpZiBuZSAuSlNPTk5hbWUgIiJ9fSBganNvbjoie3suSlNPTk5hbWV9fSJge3tlbmR9fXt7ZW5kfX0KfQp7e2VuZH19",
+    }
+}
