+++ plugins/go/openapi_go_generator/templates/client.go.tmpl
@@ -1,0 +1,60 @@
+package {{.Renderer.Package}}
+
+import (
+  "bytes"
+  "encoding/json"
+  "fmt"
+  "net/http"
+  "strings"
+)
+  
+const service = "http://localhost:8080"
+
+type Client struct {
+	Service string
+} 
+
+func NewClient(service string) *Client {
+	client := &Client{}
+	client.Service = service
+	return client
+}
+
+{{range .Renderer.Methods}}
+{{if eq .ResultTypeName ""}}
+func (client *Client) {{.ClientName}}({{parameterList .}}) (err error) {
+{{else}}
+func (client *Client) {{.ClientName}}({{parameterList .}}) (result *{{.ResultTypeName}}, err error) {
+{{end}}
+	path := client.Service + "{{.Path}}"
+	
+	{{range .ParametersType.Fields}}	
+	{{if eq .Position "path"}}
+	path = strings.Replace(path, "{" + "{{.JSONName}}" + "}", fmt.Sprintf("%v", {{.JSONName}}), 1)
+	{{end}}
+	{{end}}
+	
+	{{if eq .Method "POST"}}
+	body := new(bytes.Buffer)
+	json.NewEncoder(body).Encode({{bodyParameterName .}})
+	req, err := http.NewRequest("{{.Method}}", path, body)
+	{{else}}
+	req, err := http.NewRequest("{{.Method}}", path, nil)
+	{{end}}
+	
+	if err != nil {
+		return
+	}
+	resp, err := http.DefaultClient.Do(req)
+	if err != nil {
+		return 
+	}
+	defer resp.Body.Close()
+	{{if ne .ResultTypeName ""}}
+	decoder := json.NewDecoder(resp.Body)
+	result = &{{.ResultTypeName}}{}
+	decoder.Decode(result)
+	{{end}}
+	return  
+}
+{{end}}
