+++ plugins/go/openapi_go_generator/templates/service.go.tmpl
@@ -1,0 +1,16 @@
+package {{.Renderer.Package}}
+
+import (
+	"log"
+)
+
+func initialize_service() {
+}
+
+{{range .Renderer.Methods}}
+func {{.ProcessorName}}(parameters *{{.ParametersTypeName}}, responses *{{.ResponsesTypeName}}) (err error) {
+	log.Printf("{{.Name}} %+v", parameters)
+	// COMPLETE THIS MANUALLY
+	return nil
+}
+{{end}}
