+++ plugins/go/openapi_go_generator/templates/types.go.tmpl
@@ -1,0 +1,11 @@
+package {{.Renderer.Package}}
+
+import (
+)
+
+// types
+{{range .Renderer.Types}}
+type {{.Name}} struct { {{range .Fields}}
+  {{.Name}} {{.Type}}{{if ne .JSONName ""}} `json:"{{.JSONName}}"`{{end}}{{end}}
+}
+{{end}}
